backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
authlib==1.3.0
httpx==0.26.0
aiofiles==23.2.0

sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

beautifulsoup4==4.12.0
markdown==3.5.0
PyPDF2==3.0.0
pdfplumber==0.10.0
Pillow==10.2.0

redis==5.0.0
celery==5.3.0

python-socketio==5.11.0
websockets==12.0

boto3==1.34.0
minio==7.2.0

pytest==7.4.0
pytest-asyncio==0.23.0